Remove most exceptions -> asserts. image copy.
I was rewatching Herb Sutter's 2019 CppCon talk on exceptions and RTTI,
and was convinced by his statement of 'non-recoverable errors should be
asserted out and caught in tests'. Majority of errors in the vulkan
section of this library I'd argue are non-recoverable (some are, but
still), so, I figured exchanging 'exceptions' for 'asserts' in these
cases.